# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 工业设备智能运维系统 API 契约
  version: 2.0.0
  description: |
    本文件是成员 A（设备资产与监测）、B（健康评估与预警）、
    C（维修工单与备件）、D（身份、通知与集成）共同遵守的唯一 HTTP 契约。

    面向前端的请求统一通过本地网关 http://localhost:8080。
    服务间调用可使用各模块直连地址，并必须携带 X-Trace-Id、
    Idempotency-Key（写操作）和 X-Internal-Token。

    字段、枚举、状态码和路径的任何变更必须先修改本文件，
    再由提供方和调用方共同评审后实现。
servers:
  - url: http://localhost:8080
    description: 本地统一网关（成员 D 维护路由）
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/work-orders:
    get:
      tags: [C-维修工单]
      summary: 分页查询维修工单
      operationId: listWorkOrders
      x-contract-id: C-API-01
      x-owner-member: MEMBER_C
      x-owner-port: 8103
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
        - name: equipmentId
          in: query
          schema:
            $ref: '#/components/schemas/EquipmentId'
        - name: warningId
          in: query
          schema:
            $ref: '#/components/schemas/WarningId'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WorkOrderStatus'
        - name: assigneeId
          in: query
          schema:
            $ref: '#/components/schemas/UserId'
        - $ref: '#/components/parameters/PageParameter'
        - $ref: '#/components/parameters/PageSizeParameter'
      responses:
        '200':
          description: 按 createdAt 降序返回
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrderPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [C-维修工单]
      summary: 创建人工报修或巡检工单
      operationId: createManualWorkOrder
      x-contract-id: C-API-02
      x-owner-member: MEMBER_C
      x-owner-port: 8103
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkOrderRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/work-orders/{orderId}:
    parameters:
      - $ref: '#/components/parameters/OrderIdParameter'
    get:
      tags: [C-维修工单]
      summary: 查询工单详情
      operationId: getWorkOrderById
      x-contract-id: C-API-03
      x-owner-member: MEMBER_C
      x-owner-port: 8103
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrder'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/work-orders/{orderId}/commands:
    parameters:
      - $ref: '#/components/parameters/OrderIdParameter'
    post:
      tags: [C-维修工单]
      summary: 执行工单状态命令
      description: |
        使用 action 驱动状态迁移。各 action 的允许源状态、目标状态、
        必填扩展字段和角色见 docs/06_详细接口约定与联调手册.md。
      operationId: executeWorkOrderCommand
      x-contract-id: C-API-04
      x-owner-member: MEMBER_C
      x-owner-port: 8103
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkOrderCommandRequest'
      responses:
        '200':
          description: 命令成功；返回迁移后的完整工单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/spare-parts:
    get:
      tags: [C-维修工单]
      summary: 分页查询备件和可用库存
      operationId: listSpareParts
      x-contract-id: C-API-05
      x-owner-member: MEMBER_C
      x-owner-port: 8103
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
        - name: keyword
          in: query
          schema:
            type: string
            maxLength: 50
        - name: lowStockOnly
          in: query
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/PageParameter'
        - $ref: '#/components/parameters/PageSizeParameter'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SparePartPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/work-orders/{orderId}/spare-requests:
    parameters:
      - $ref: '#/components/parameters/OrderIdParameter'
    post:
      tags: [C-维修工单]
      summary: 为工单提交备件申请
      operationId: createSpareRequest
      x-contract-id: C-API-06
      x-owner-member: MEMBER_C
      x-owner-port: 8103
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpareRequestRequest'
      responses:
        '201':
          description: 申请创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpareRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/spare-requests/{requestId}/commands:
    parameters:
      - $ref: '#/components/parameters/SpareRequestIdParameter'
    post:
      tags: [C-维修工单]
      summary: 审批、预留、领用、退回或取消备件申请
      operationId: executeSpareRequestCommand
      x-contract-id: C-API-07
      x-owner-member: MEMBER_C
      x-owner-port: 8103
      parameters:
        - $ref: '#/components/parameters/TraceIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpareRequestCommandRequest'
      responses:
        '200':
          description: 命令成功；返回更新后的申请
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpareRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录接口签发的访问令牌
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: 本地开发由 INTERNAL_API_TOKEN 环境变量提供
  parameters:
    TraceIdHeader:
      name: X-Trace-Id
      in: header
      required: false
      description: 调用方可传；缺失时网关生成并回传
      schema:
        type: string
        minLength: 8
        maxLength: 64
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID；同一业务写请求重试时保持不变
      schema:
        type: string
        format: uuid
    PageParameter:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParameter:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    OrderIdParameter:
      name: orderId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/OrderId'
    SpareRequestIdParameter:
      name: requestId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/SpareRequestId'
...
  responses:
...
    BadRequest:
      description: 字段缺失、格式错误、枚举错误或查询参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未登录、JWT 无效或内部令牌无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 身份有效但缺少该操作所需权限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 指定业务对象不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 当前状态不允许操作、版本冲突或幂等键对应不同请求
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 服务内部错误；不得向客户端暴露堆栈或数据库细节
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    EquipmentId:
      type: string
      pattern: '^EQ-[0-9]{6}$'
      example: EQ-000001
    WarningId:
      type: string
      pattern: '^WARN-[0-9]{8}-[0-9]{4}$'
      example: WARN-20260917-0001
    OrderId:
      type: string
      pattern: '^WO-[0-9]{8}-[0-9]{4}$'
      example: WO-20260917-0001
    SparePartId:
      type: string
      pattern: '^SP-[0-9]{6}$'
      example: SP-000001
    SpareRequestId:
      type: string
      pattern: '^SPR-[0-9]{8}-[0-9]{4}$'
      example: SPR-20260917-0001
    UserId:
      type: string
      pattern: '^USER-[A-Z0-9-]{1,27}$'
      example: USER-C-002
...
    WorkOrderStatus:
      type: string
      enum:
        - PENDING_CONFIRMATION
        - PENDING_ASSIGNMENT
        - PENDING_ACCEPTANCE
        - PENDING_MAINTENANCE
        - MAINTAINING
        - WAITING_PARTS
        - PENDING_INSPECTION
        - COMPLETED
        - CANCELLED
    WorkOrderPriority:
      type: string
      enum: [P1, P2, P3, P4]
    WorkOrderSource:
      type: string
      enum: [WARNING, MANUAL, INSPECTION]
    WorkOrderAction:
      type: string
      enum:
        - CONFIRM
        - ASSIGN
        - ACCEPT
        - START
        - WAIT_FOR_PARTS
        - RESUME
        - SUBMIT_FOR_INSPECTION
        - PASS_INSPECTION
        - REJECT_INSPECTION
        - CANCEL
    SpareRequestStatus:
      type: string
      enum:
        - PENDING_APPROVAL
        - APPROVED
        - RESERVED
        - ISSUED
        - PARTIALLY_RETURNED
        - CLOSED
        - REJECTED
        - CANCELLED
    SpareRequestAction:
      type: string
      enum: [APPROVE, REJECT, RESERVE, ISSUE, RETURN, CLOSE, CANCEL]
    MaintenanceResult:
      type: string
      enum: [RECOVERED, PARTIALLY_RECOVERED, NOT_RECOVERED]
...
    WorkOrder:
      type: object
      required:
        - orderId
        - sourceType
        - warningId
        - equipmentId
        - equipmentNameSnapshot
        - title
        - description
        - priority
        - status
        - reporterId
        - assigneeId
        - dueAt
        - conclusion
        - version
        - createdAt
        - updatedAt
      properties:
        orderId:
          $ref: '#/components/schemas/OrderId'
        sourceType:
          $ref: '#/components/schemas/WorkOrderSource'
        warningId:
          oneOf:
            - $ref: '#/components/schemas/WarningId'
            - type: 'null'
        equipmentId:
          $ref: '#/components/schemas/EquipmentId'
        equipmentNameSnapshot:
          type: string
          minLength: 1
          maxLength: 100
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          minLength: 1
          maxLength: 2000
        priority:
          $ref: '#/components/schemas/WorkOrderPriority'
        status:
          $ref: '#/components/schemas/WorkOrderStatus'
        reporterId:
          $ref: '#/components/schemas/UserId'
        assigneeId:
          oneOf:
            - $ref: '#/components/schemas/UserId'
            - type: 'null'
        dueAt:
          type: [string, 'null']
          format: date-time
        conclusion:
          oneOf:
            - $ref: '#/components/schemas/MaintenanceConclusion'
            - type: 'null'
        version:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    WorkOrderPage:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkOrder'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
    CreateWorkOrderRequest:
      type: object
      required:
        - sourceType
        - equipmentId
        - title
        - description
        - priority
        - reporterId
      properties:
        sourceType:
          type: string
          enum: [MANUAL, INSPECTION]
        equipmentId:
          $ref: '#/components/schemas/EquipmentId'
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          minLength: 1
          maxLength: 2000
        priority:
          $ref: '#/components/schemas/WorkOrderPriority'
        reporterId:
          $ref: '#/components/schemas/UserId'
        dueAt:
          type: [string, 'null']
          format: date-time
      additionalProperties: false
    WorkOrderCommandRequest:
      type: object
      required: [action, operatorId, expectedVersion]
      properties:
        action:
          $ref: '#/components/schemas/WorkOrderAction'
        operatorId:
          $ref: '#/components/schemas/UserId'
        expectedVersion:
          type: integer
          minimum: 0
        assigneeId:
          oneOf:
            - $ref: '#/components/schemas/UserId'
            - type: 'null'
          description: action=ASSIGN 时必填
        comment:
          type: [string, 'null']
          maxLength: 500
        conclusion:
          oneOf:
            - $ref: '#/components/schemas/MaintenanceConclusionInput'
            - type: 'null'
          description: action=SUBMIT_FOR_INSPECTION 时必填
      additionalProperties: false
    MaintenanceConclusionInput:
      type: object
      required: [rootCause, measures, result, effective, completedAt]
      properties:
        rootCause:
          type: string
          minLength: 1
          maxLength: 500
        measures:
          type: string
          minLength: 1
          maxLength: 1000
        result:
          $ref: '#/components/schemas/MaintenanceResult'
        effective:
          type: boolean
        completedAt:
          type: string
          format: date-time
        downtimeMinutes:
          type: [integer, 'null']
          minimum: 0
      additionalProperties: false
    MaintenanceConclusion:
      allOf:
        - $ref: '#/components/schemas/MaintenanceConclusionInput'
        - type: object
          required: [submittedBy]
          properties:
            submittedBy:
              $ref: '#/components/schemas/UserId'
    SparePart:
      type: object
      required:
        - sparePartId
        - name
        - specification
        - unit
        - onHandQuantity
        - reservedQuantity
        - availableQuantity
        - reorderPoint
        - version
      properties:
        sparePartId:
          $ref: '#/components/schemas/SparePartId'
        name:
          type: string
          minLength: 1
          maxLength: 100
        specification:
          type: string
          minLength: 1
          maxLength: 200
        unit:
          type: string
          minLength: 1
          maxLength: 20
          example: 个
        onHandQuantity:
          type: integer
          minimum: 0
        reservedQuantity:
          type: integer
          minimum: 0
        availableQuantity:
          type: integer
          minimum: 0
          description: onHandQuantity - reservedQuantity
        reorderPoint:
          type: integer
          minimum: 0
        version:
          type: integer
          minimum: 0
      additionalProperties: false
    SparePartPage:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SparePart'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
    SpareRequest:
      type: object
      required:
        - requestId
        - orderId
        - sparePartId
        - requestedQuantity
        - approvedQuantity
        - issuedQuantity
        - returnedQuantity
        - status
        - requesterId
        - approverId
        - version
        - createdAt
        - updatedAt
      properties:
        requestId:
          $ref: '#/components/schemas/SpareRequestId'
        orderId:
          $ref: '#/components/schemas/OrderId'
        sparePartId:
          $ref: '#/components/schemas/SparePartId'
        requestedQuantity:
          type: integer
          minimum: 1
        approvedQuantity:
          type: integer
          minimum: 0
        issuedQuantity:
          type: integer
          minimum: 0
        returnedQuantity:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/SpareRequestStatus'
        requesterId:
          $ref: '#/components/schemas/UserId'
        approverId:
          oneOf:
            - $ref: '#/components/schemas/UserId'
            - type: 'null'
        reason:
          type: [string, 'null']
          maxLength: 500
        version:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    CreateSpareRequestRequest:
      type: object
      required: [sparePartId, requestedQuantity, requesterId, reason]
      properties:
        sparePartId:
          $ref: '#/components/schemas/SparePartId'
        requestedQuantity:
          type: integer
          minimum: 1
          maximum: 10000
        requesterId:
          $ref: '#/components/schemas/UserId'
        reason:
          type: string
          minLength: 1
          maxLength: 500
      additionalProperties: false
    SpareRequestCommandRequest:
      type: object
      required: [action, operatorId, expectedVersion]
      properties:
        action:
          $ref: '#/components/schemas/SpareRequestAction'
        operatorId:
          $ref: '#/components/schemas/UserId'
        expectedVersion:
          type: integer
          minimum: 0
        quantity:
          type: [integer, 'null']
          minimum: 1
          description: APPROVE、ISSUE、RETURN 时按动作规则填写
        comment:
          type: [string, 'null']
          maxLength: 500
      additionalProperties: false
...
    FieldViolation:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
      additionalProperties: false
    ErrorResponse:
      type: object
      required: [code, message, traceId, timestamp, details]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
          example: EQUIPMENT_NOT_FOUND
        message:
          type: string
          example: 设备不存在
        traceId:
          type: string
          minLength: 8
          maxLength: 64
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'
      additionalProperties: false